refactor(samplegen): Move files out of core/tests into samplegen
For AB#1431
