+++ YoFi.SampleGen/YoFi.SampleGen.Exe/YoFi.SampleGen.Exe.csproj
@@ -3,8 +3,10 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\YoFi.SampleGen\YoFi.SampleGen.csproj" />
+  </ItemGroup>
 
 </Project>
