+++ YoFi.SampleGen/YoFi.SampleGen.Tests/GeneratorDefinitionTest.cs
