+++ YoFi.SampleGen/YoFi.SampleGen.Tests/Helpers/SampleData.cs
@@ -1,0 +1,28 @@
+﻿using System.IO;
+using System.Linq;
+using System.Reflection;
+
+namespace Common.DotNet.Test
+{
+    public static class SampleData
+    {
+        public static Stream Open(string filename)
+        {
+            var assy = Assembly.GetExecutingAssembly();
+            var names = assy.GetManifestResourceNames();
+
+            var matching = names.Where(x => x.Contains(filename));
+            if (!matching.Any())
+                throw new FileNotFoundException($"{filename} not found in assembly");
+            if (matching.Skip(1).Any())
+                throw new FileNotFoundException($"{filename} is ambiguous. Multiple found in assembly");
+
+            var name = matching.First();
+            var stream = assy.GetManifestResourceStream(name);
+            if (null == stream)
+                throw new FileNotFoundException($"{filename} not found in assembly");
+
+            return stream;
+        }
+    }
+}
