--- YoFi.SampleGen/YoFi.SampleGen.Tests/UnitTest1.cs
@@ -1,12 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace YoFi.SampleGen.Tests;
-
-[TestClass]
-public class UnitTest1
-{
-    [TestMethod]
-    public void TestMethod1()
-    {
-    }
-}
