--- YoFi.SampleGen/YoFi.SampleGen/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace YoFi.SampleGen;
-public class Class1
-{
-
-}
