+++ YoFi.SampleGen/YoFi.SampleGen/SampleDataPattern.cs
