+++ YoFi.SampleGen/YoFi.SampleGen/YoFi.SampleGen.csproj
@@ -2,8 +2,10 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\YoFi.Core\YoFi.Core.csproj" />
+  </ItemGroup>
 
 </Project>
