+++ YoFi.Tests/Controllers/TransactionControllerSlimTest.cs
@@ -12,7 +12,7 @@ using YoFi.Core;
 using YoFi.Core.Models;
 using YoFi.Core.Repositories;
 using YoFi.Core.Repositories.Wire;
-using YoFi.Core.SampleGen;
+using YoFi.Core.SampleData;
 using YoFi.Tests.Helpers;
 
 namespace YoFi.Tests.Controllers.Slim
